+++ src/main/java/org/bnjax3/noitacraft/spell/main_classes/MaterialSpell.java
@@ -6,7 +6,7 @@ public class MaterialSpell extends Spell{
     public final Block material;
     public final boolean replace;
     public final int radius;
-    public MaterialSpell(int uses, int manaDrain, int castDelay, int rechargeTime, float spread, float recoil, Block material, boolean replace, int radius) {
+    public MaterialSpell(int uses, int manaDrain, float castDelay, float rechargeTime, float spread, float recoil, Block material, boolean replace, int radius) {
         super(uses, manaDrain, castDelay, rechargeTime, spread, recoil, true);
         this.material = material;
         this.replace = replace;
